Financial_Sentiment_System/etl_pipeline/requirements.txt
pymongo
yfinance
requests
textblob
python-dotenv